+++ src/errors/index.ts
@@ -1,6 +1,7 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
 export { BaseError } from './BaseError';
+export { checkBaseInputTypes } from './checkBaseInputTypes';
 export { checkLocalTxInput } from './checkLocalTxInputs';
 export { checkXcmTxInputs } from './checkXcmTxInputs';
 export { checkXcmVersion } from './checkXcmVersion';