+++ src/testHelpers/adjustedMockSystemApi.ts
@@ -1,9 +1,12 @@
 // Copyright 2023 Parity Technologies (UK) Ltd.
 
 import type { ApiPromise } from '@polkadot/api';
+import { Compact } from '@polkadot/types';
+import { BlockNumber } from '@polkadot/types/interfaces';
 
 import { mockSystemApi } from './mockSystemApi';
 import { mockWeightInfo } from './mockWeightInfo';
+import { statemineV9420 } from './metadata/statemineV9420';
 
 const getSystemRuntimeVersion = () =>
 	Promise.resolve().then(() => {
@@ -23,6 +26,17 @@ const queryInfoCallAt = () =>
 		mockSystemApi.createType('RuntimeDispatchInfoV2', mockWeightInfo)
 	);
 
+const getMetadata = () => Promise.resolve().then(() => mockSystemApi.registry.createType('Metadata', statemineV9420));
+
+const getHeader = (): Promise<{ number: Compact<BlockNumber> }> =>
+	Promise.resolve().then(() => {
+		return {
+			number: mockSystemApi.registry.createType('Compact<BlockNumber>', 100),
+		};
+	});
+
+const createType = mockSystemApi.registry.createType.bind(mockSystemApi);
+
 const mockApiAt = {
 	call: {
 		transactionPaymentApi: {
@@ -32,11 +46,16 @@ const mockApiAt = {
 };
 
 export const adjustedMockSystemApi = {
+	createType: createType,
 	registry: mockSystemApi.registry,
 	rpc: {
 		state: {
 			getRuntimeVersion: getSystemRuntimeVersion,
+			getMetadata: getMetadata
 		},
+		chain: {
+			getHeader: getHeader
+		}
 	},
 	query: {
 		polkadotXcm: {